infrastructure/requirements.txt
pulumi>=2.0.0,<3.0.0
pulumi-aws>=2.0.0,<3.0.0
python-dotenv<=0.13.0
ipaddress
